This commit contains the modifications regarding a consistent implementation with the copy of the configuration file parallel.
Previously, the files were sequentially copied and the copy
instructions called `os.system` besides being fragmented
along the file. The new approach adopted implements a more
idiomatic form to treat desktop environments and a lazy copy
process, that is, in the construction of the object, based on
the user's and the preferences of the user, the set of files to be
copied is created to, only with the call of the `run` method, to be
performed in parallel.

In order to increase the maintenance and legibilization of the code,
the attributions were properly separated and the type notes, in most
cases were created.

The original attributions of this file were divided into:

- Prepare the environment;
- Copy;
- Collaborative actions;

The environment preparation actions have been moved to the `setup`
method which, thanks to 'SandwichMeta', are executed before `run`,
`run` which contains the actions to be executed - it is indicated
to execute '__parallel_copy' within the method -, the copy actions
have been abstracted in '__parallel_copy', the changes after 'run'
must be in 'teardown'.